[FIX] move, pass-by-ref
